Consolidate node stores: sharedNodeStore as single source of truth (#679)
## Summary
- Converted sharedNodeStore to Svelte 5 $state for automatic reactivity
- Migrated TextNode, DateNode from nodeData to sharedNodeStore
- Updated OptimisticOperationManager to use sharedNodeStore
- Removed deprecated reactive-node-data.svelte.ts (and its tests)
- Removed viewedNodeCache workaround (no longer needed with $state)
- Updated all imports to use .svelte.ts extension

## Changes
- `sharedNodeStore.nodes` now uses `$state(new Map<string, Node>())`
- Added `snapshot()` and `restore()` methods for optimistic rollback
- Components using `$derived(sharedNodeStore.getNode(nodeId))` now
  automatically re-render when nodes change (cross-pane reactivity)

## Test Results
- Frontend: 74 test files, 1469 tests passed (24 tests removed with deleted file)
- Backend: 516 passed, 0 failed
