+++ packages/desktop-app/src/lib/components/text-node.svelte
@@ -10,7 +10,7 @@
 <script lang="ts">
   import { createEventDispatcher } from 'svelte';
   import BaseNode from '$lib/design/components/base-node.svelte';
-  import { nodeData } from '$lib/stores/reactive-node-data.svelte';
+  import { sharedNodeStore } from '$lib/services/shared-node-store.svelte';
   import { structureTree } from '$lib/stores/reactive-structure-tree.svelte';
 
   const dispatch = createEventDispatcher();
@@ -30,9 +30,10 @@
     children?: string[];
   } = $props();
 
-  // Use reactive stores directly instead of relying on props
-  // Components query the stores for current data and re-render automatically when data changes
-  let node = $derived(nodeData.getNode(nodeId));
+  // Use sharedNodeStore as single source of truth for cross-pane reactivity
+  // This ensures content changes from other panes are immediately reflected
+  // Issue #679: Migrated from nodeData (which was never receiving updates)
+  let node = $derived(sharedNodeStore.getNode(nodeId));
   let childIds = $derived(structureTree.getChildren(nodeId));
 
   // Derive props from stores with fallback to passed props for backward compatibility
