+++ packages/desktop-app/src/tests/services/node-type-conflict-detection.test.ts
@@ -10,7 +10,7 @@
  */
 
 import { describe, it, expect, beforeEach, afterEach } from 'vitest';
-import { SharedNodeStore } from '../../lib/services/shared-node-store';
+import { SharedNodeStore } from '../../lib/services/shared-node-store.svelte';
 import type { Node } from '../../lib/types';
 import type { UpdateSource } from '../../lib/types/update-protocol';
